ETL/3_process_data/requirements.txt
functions-framework==3.*
google-cloud-storage
pandas
pyarrow
fastparquet
